@@ -15,19 +15,25 @@
     "@types/mapbox-gl": "^2.3.0",
     "@types/socket.io": "^3.0.2",
     "@types/uniqid": "^5.3.0",
+    "buffer": "^6.0.3",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "geolib": "^3.3.1",
+    "hyntax": "^1.1.9",
     "mapbox-gl": "^2.3.0",
     "mithril": "^2.0.4",
     "node-fetch": "^2.6.1",
+    "npm": "^7.20.3",
     "path": "^0.12.7",
+    "querystring": "^0.2.1",
     "redux-logger": "^3.0.6",
     "rxjs": "^6.6.7",
     "socket.io": "^4.1.2",
     "socket.io-client": "^4.1.2",
-    "uniqid": "^5.3.0"
+    "stream-browserify": "^3.0.0",
+    "uniqid": "^5.3.0",
+    "wikijs": "^6.2.3"
   },
   "devDependencies": {
     "@babel/core": "^7.14.6",
@@ -42,7 +48,7 @@
     "concurrently": "^6.2.0",
     "cross-env": "^7.0.3",
     "css-loader": "^5.2.6",
-    "eslint": "^7.29.0",
+    "eslint": "^7.31.0",
     "eslint-loader": "^4.0.2",
     "extract-loader": "^5.1.0",
     "handlebars-loader": "^1.7.1",
