@@ -1,7 +1,9 @@
 import {BaseState, BaseAttrs} from "base/types"
 
 export interface GamePageState extends BaseState {
-
+    showRoundEndModal: boolean;
 }
 
-export interface GamePageAttrs extends BaseAttrs {}
+export interface GamePageAttrs extends BaseAttrs {
+
+}
