@@ -1,5 +1,6 @@
 import { combineReducers } from "redux";
 import { GameDataReducer } from "./GameData/reducer";
 import { UserDataReducer } from "./UserData/reducer";
+import { CurrentPageDataReducer } from "./currentPageData/reducer";
 
-export const rootReducer = combineReducers({UserData: UserDataReducer, GameData: GameDataReducer});
+export const rootReducer = combineReducers({UserData: UserDataReducer, GameData: GameDataReducer, CurrentPageData: CurrentPageDataReducer});
