@@ -1,4 +1,4 @@
-import { Store, createStore, applyMiddleware } from "redux";
+import { createStore, applyMiddleware } from "redux";
 import logger from 'redux-logger';
 import { rootReducer } from "./root";
 export type storeType = typeof store;
